nwg_exceptions.h
 - add ErrnoException struct
 - move error_description to nwg_exceptions.h
 - replace existing usages of runtime_error{ error_description }
   with ErrnoException
